+++ test/NonSiloTests/Orleans.NonSiloTests.csproj
@@ -190,10 +190,6 @@
       <Project>{6ff2004c-cdf8-479c-bf27-c6bfe8ef93e0}</Project>
       <Name>OrleansRuntime</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\src\OrleansServiceBus\OrleansServiceBus.csproj">
-      <Project>{8BFF0092-15D2-4425-80C0-29E381702F2B}</Project>
-      <Name>OrleansServiceBus</Name>
-    </ProjectReference>
     <ProjectReference Include="..\..\src\OrleansTestingHost\OrleansTestingHost.csproj">
       <Project>{40ee3b00-d381-485f-9c69-ff706837deed}</Project>
       <Name>OrleansTestingHost</Name>
