+++ test/Tester/GrainStateContainingGrainReferences.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Orleans.Runtime;
+
+namespace Tester
+{
+    [Serializable]
+    public class GrainStateContainingGrainReferences
+    {
+        public IAddressable Grain { get; set; }
+        public List<IAddressable> GrainList { get; set; }
+        public Dictionary<string, IAddressable> GrainDict { get; set; }
+
+        public GrainStateContainingGrainReferences()
+        {
+            GrainList = new List<IAddressable>();
+            GrainDict = new Dictionary<string, IAddressable>();
+        }
+    }
+}
