+++ test/Tester/Tester.csproj
@@ -45,6 +45,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CancellationTests\GrainCancellationTokenTests.cs" />
+    <Compile Include="GrainStateContainingGrainReferences.cs" />
     <Compile Include="ClientConnectionTests\ClientDisconnectionEventTests.cs" />
     <Compile Include="CollectionFixtures.cs" />
     <Compile Include="EventSourcingTests\AccountGrainTests.cs" />
