+++ vNext/test/NonSiloTests/Orleans.NonSiloTests.csproj
@@ -41,111 +41,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\..\test\NonSiloTests\ConfigTests.cs">
-      <Link>ConfigTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\CollectionFixtures.cs">
-      <Link>CollectionFixtures.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\AsynchAgentRestartTest.cs">
-      <Link>AsynchAgentRestartTest.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\AsyncPipelineTests.cs">
-      <Link>AsyncPipelineTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\Async_AsyncExecutorWithRetriesTests.cs">
-      <Link>Async_AsyncExecutorWithRetriesTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\Async_TimingTests.cs">
-      <Link>Async_TimingTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\CodeGeneratorTests.cs">
-      <Link>CodeGeneratorTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\CounterTests.cs">
-      <Link>CounterTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\CounterStatisticTests.cs">
-      <Link>General\CounterStatisticTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\Identifiertests.cs">
-      <Link>General\Identifiertests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\InterfaceRules.cs">
-      <Link>General\InterfaceRules.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\InternalSerializationTests.cs">
-      <Link>General\InternalSerializationTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\LruTest.cs">
-      <Link>General\LruTest.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\RequestContextTestsNonSiloRequired.cs">
-      <Link>General\RequestContextTestsNonSiloRequired.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\General\RingTests_Standalone.cs">
-      <Link>General\RingTests_Standalone.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\ManagementAgentTests.cs">
-      <Link>ManagementAgentTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\OrleansRuntime\AsyncSerialExecutorTests.cs">
-      <Link>OrleansRuntime\AsyncSerialExecutorTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\OrleansRuntime\ExceptionsTests.cs">
-      <Link>OrleansRuntime\ExceptionsTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\OrleansRuntime\Streams\BestFitBalancerTests.cs">
-      <Link>OrleansRuntime\Streams\BestFitBalancerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\OrleansRuntime\Streams\SubscriptionMarkerTests.cs">
-      <Link>OrleansRuntime\Streams\SubscriptionMarkerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\Serialization\BuiltInSerializerTests.cs">
-      <Link>Serialization\BuiltInSerializerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\Serialization\ILBasedExceptionSerializerTests.cs">
-      <Link>Serialization\ILBasedExceptionSerializerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\Serialization\ILBasedSerializerTests.cs">
-      <Link>Serialization\ILBasedSerializerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\Serialization\SerializerGenerationTests.cs">
-      <Link>Serialization\SerializerGenerationTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SchedulerTests\OrleansTaskSchedulerAdvancedTests.cs">
-      <Link>SchedulerTests\OrleansTaskSchedulerAdvancedTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SchedulerTests\OrleansTaskSchedulerAdvancedTests_Set2.cs">
-      <Link>SchedulerTests\OrleansTaskSchedulerAdvancedTests_Set2.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SchedulerTests\OrleansTaskSchedulerBasicTests.cs">
-      <Link>SchedulerTests\OrleansTaskSchedulerBasicTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializationTests\ConfigurationTests\SerializationProviderTests.cs">
-      <Link>SerializationTests\ConfigurationTests\SerializationProviderTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializationTests\ExternalSerializerTest.cs">
-      <Link>SerializationTests\ExternalSerializerTest.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializationTests\SerializationOrderTests.cs">
-      <Link>SerializationTests\SerializationOrderTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializationTests\SerializationTests.DifferentTypes.cs">
-      <Link>SerializationTests\SerializationTests.DifferentTypes.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializationTests\SerializationTests.ImmutableCollections.cs">
-      <Link>SerializationTests\SerializationTests.ImmutableCollections.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializerTests\CustomSerializerTests.cs">
-      <Link>SerializerTests\CustomSerializerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\SerializerTests\MessageSerializerTests.cs">
-      <Link>SerializerTests\MessageSerializerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\NonSiloTests\TestInternalHelper.cs">
-      <Link>TestInternalHelper.cs</Link>
-    </Compile>
+    <Compile Include="..\..\..\test\NonSiloTests\**\*.cs" Exclude="
+    ..\..\..\test\NonSiloTests\obj\**\*.cs;
+    ..\..\..\test\NonSiloTests\bin\**\*.cs;
+    ..\..\..\test\NonSiloTests\Properties\*.cs;
+    ..\..\..\test\NonSiloTests\AssemblyLoaderTests.cs" />
   </ItemGroup>
 
   <ItemGroup>
