+++ vNext/test/TestInternalGrainInterface/Properties/AssemblyInfo.cs
@@ -19,6 +19,7 @@ using Orleans.CodeGeneration;
 
 [assembly: InternalsVisibleTo("TestInternalGrains")]
 [assembly: InternalsVisibleTo("Tester")]
+[assembly: InternalsVisibleTo("TesterInternal")]
 [assembly: InternalsVisibleTo("DefaultCLuster.Tests")]
 
 // generate Orleans serializers for types in FSharp.core.dll
