+++ vNext/test/TesterAzureUtils/Tester.AzureUtils.csproj
@@ -47,70 +47,11 @@
     <ProjectReference Include="..\TestInternalGrains\TestInternalGrains.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureGossipTableTests.cs">
-      <Link>AzureGossipTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\CollectionFixtures.cs">
-      <Link>CollectionFixtures.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureMembershipTableTests.cs">
-      <Link>AzureMembershipTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureQueueDataManagerTests.cs">
-      <Link>AzureQueueDataManagerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureRemindersTableTests.cs">
-      <Link>AzureRemindersTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureStorageBasicTestFixture.cs">
-      <Link>AzureStorageBasicTestFixture.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureTableDataManagerStressTests.cs">
-      <Link>AzureTableDataManagerStressTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureTableDataManagerTests.cs">
-      <Link>AzureTableDataManagerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\AzureTableErrorCodeTests.cs">
-      <Link>AzureTableErrorCodeTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Deployment\AzureSiloTests.cs">
-      <Link>Deployment\AzureSiloTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\GenericGrainsInAzureStorageTests.cs">
-      <Link>GenericGrainsInAzureStorageTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Persistence\PersistenceProviderTests.cs">
-      <Link>Persistence\PersistenceProviderTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\SiloInstanceTableManagerTests.cs">
-      <Link>SiloInstanceTableManagerTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\AQClientStreamTests.cs">
-      <Link>Streaming\AQClientStreamTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\AQStreamFilteringTests.cs">
-      <Link>Streaming\AQStreamFilteringTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\AQStreamingTests.cs">
-      <Link>Streaming\AQStreamingTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\AQSubscriptionMultiplicityTests.cs">
-      <Link>Streaming\AQSubscriptionMultiplicityTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\AzureQueueAdapterTests.cs">
-      <Link>Streaming\AzureQueueAdapterTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\SampleAzureQueueStreamingTests.cs">
-      <Link>Streaming\SampleAzureQueueStreamingTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\Streaming\TestAzureQueueStreamProvider.cs">
-      <Link>Streaming\TestAzureQueueStreamProvider.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterAzureUtils\UnitTestAzureTableDataManager.cs">
-      <Link>UnitTestAzureTableDataManager.cs</Link>
-    </Compile>
+    <ItemGroup>
+    <Compile Include="..\..\..\test\TesterAzureUtils\**\*.cs" Exclude="
+    ..\..\..\test\TesterAzureUtils\obj\**\*.cs;
+    ..\..\..\test\TesterAzureUtils\bin\**\*.cs;
+    ..\..\..\test\TesterAzureUtils\Properties\*.cs" />
   </ItemGroup>
 
   <ItemGroup>
