+++ vNext/test/TesterInternal/TesterInternal.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup Label="Configuration">
     <DefineConstants>NETSTANDARD;NETSTANDARD_TODO</DefineConstants>
   </PropertyGroup>
@@ -56,96 +56,17 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\..\test\TesterInternal\TestUtils.cs">
-      <Link>TestUtils.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\CollectionFixtures.cs">
-      <Link>CollectionFixtures.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\GatewaySelectionTest.cs">
-      <Link>GatewaySelectionTest.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\TestStoreGrainState.cs">
-      <Link>TestStoreGrainState.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\ConnectionStringFixture.cs">
-      <Link>ConnectionStringFixture.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\ErrorInjectionStorageProvider.cs">
-      <Link>ErrorInjectionStorageProvider.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\MembershipTests\MembershipTableTestsBase.cs">
-      <Link>MembershipTests\MembershipTableTestsBase.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\RemindersTest\ReminderTableTestsBase.cs">
-      <Link>ReminderTests\ReminderTableTestsBase.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\TimerTests\ReminderTests_Base.cs">
-      <Link>TimerTests\ReminderTests_Base.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\CommonStorageTests.cs">
-      <Link>StorageTests\CommonStorageTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\CommonStorageUtilities.cs">
-      <Link>StorageTests\CommonStorageUtilities.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\HierarchicalKeyStoreTests.cs">
-      <Link>StorageTests\HierarchicalKeyStoreTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\LocalStoreTests.cs">
-      <Link>StorageTests\LocalStoreTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\RandomUtilities.cs">
-      <Link>StorageTests\RandomUtilities.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\Range.cs">
-      <Link>StorageTests\Range.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\SymbolSet.cs">
-      <Link>StorageTests\SymbolSet.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\GrainTypeGenerator.cs">
-      <Link>StorageTests\TestDataSets\GrainTypeGenerator.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\StorageDataSet2CyrillicIdsAndGrainNames.cs">
-      <Link>StorageTests\TestDataSets\StorageDataSet2CyrillicIdsAndGrainNames.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\StorageDataSetGeneric.cs">
-      <Link>StorageTests\TestDataSets\StorageDataSetGeneric.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\StorageDataSetPlain.cs">
-      <Link>StorageTests\TestDataSets\StorageDataSetPlain.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\StorageDateSetGenericHuge.cs">
-      <Link>StorageTests\TestDataSets\StorageDateSetGenericHuge.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\TestState1.cs">
-      <Link>StorageTests\TestDataSets\TestState1.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\TestDataSets\TestStateGeneric1.cs">
-      <Link>StorageTests\TestDataSets\TestStateGeneric1.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\StorageProviders\BaseJSONStorageProvider.cs">
-      <Link>StorageTests\StorageProvidesr\BaseJSONStorageProvider.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\StorageProviders\FileStorageProvider.cs">
-      <Link>StorageTests\StorageProvidesr\FileStorageProvider.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StorageTests\StorageProviders\IJSONStateDataManager.cs">
-      <Link>StorageTests\StorageProvidesr\IJSONStateDataManager.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StreamingTests\MultipleStreamsTestRunner.cs">
-      <Link>StreamingTests\MultipleStreamsTestRunner.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StreamingTests\SingleStreamTestRunner.cs">
-      <Link>StreamingTests\SingleStreamTestRunner.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StreamingTests\StreamTestHelperClasses.cs">
-      <Link>StreamingTests\StreamTestHelperClasses.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterInternal\StreamingTests\StreamTestUtils.cs">
-      <Link>StreamingTests\StreamTestUtils.cs</Link>
-    </Compile>
+    <Compile Include="..\..\..\test\TesterInternal\**\*.cs" Exclude="
+    ..\..\..\test\TesterInternal\obj\**\*.cs;
+    ..\..\..\test\TesterInternal\bin\**\*.cs;
+    ..\..\..\test\TesterInternal\Properties\*.cs;
+    ..\..\..\test\TesterInternal\CounterControlProgTests.cs;
+    ..\..\..\test\TesterInternal\GeoClusterTests\*.cs;
+    ..\..\..\test\TesterInternal\HostedTestClusterBase.cs;
+    ..\..\..\test\TesterInternal\OrleansHostProgTests.cs;
+    ..\..\..\test\TesterInternal\General\BootstrapProviderTests.cs;
+    ..\..\..\test\TesterInternal\StorageTests\PersistenceGrainTests.cs;
+    ..\..\..\test\TesterInternal\StreamingTests\StreamPubSubReliabilityTests.cs" />
   </ItemGroup>
 
   <ItemGroup>
