+++ vNext/test/TesterSQLUtils/Tester.SQLUtils.csproj
@@ -47,97 +47,11 @@
     <ProjectReference Include="..\TestInternalGrainInterface\TestInternalGrainInterfaces.csproj" />
     <ProjectReference Include="..\TestInternalGrains\TestInternalGrains.csproj" />
   </ItemGroup>
-  <ItemGroup>
-    <Compile Include="..\..\..\test\TesterSQLUtils\CollectionFixtures.cs">
-      <Link>CollectionFixtures.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\RelationalUtilities\MySqlStorageForTesting.cs">
-      <Link>RelationalUtilities\MySqlStorageForTesting.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\RelationalUtilities\PostgreSqlStorageForTesting.cs">
-      <Link>RelationalUtilities\PostgreSqlStorageForTesting.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\RelationalUtilities\RelationalStorageForTesting.cs">
-      <Link>RelationalUtilities\RelationalStorageForTesting.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\RelationalUtilities\SqlServerStorageForTesting.cs">
-      <Link>RelationalUtilities\SqlServerStorageForTesting.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\DummyCounter.cs">
-      <Link>SqlStatisticsPublisherTests\DummyCounter.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\DummyPerformanceMetrics.cs">
-      <Link>SqlStatisticsPublisherTests\DummyPerformanceMetrics.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\MySqlStatisticsPublisherTests.cs">
-      <Link>SqlStatisticsPublisherTests\MySqlStatisticsPublisherTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\PostgreSqlStatisticsPublisherTests.cs">
-      <Link>SqlStatisticsPublisherTests\PostgreSqlStatisticsPublisherTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\SqlServerStatisticsPublisherTests.cs">
-      <Link>SqlStatisticsPublisherTests\SqlServerStatisticsPublisherTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\SqlStatisticsPublisherTestsBase.cs">
-      <Link>SqlStatisticsPublisherTests\SqlStatisticsPublisherTestsBase.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\StatisticsPublisherProviderConfig.cs">
-      <Link>SqlStatisticsPublisherTests\StatisticsPublisherProviderConfig.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlStatisticsPublisherTests\StatisticsPublisherProviderRuntime.cs">
-      <Link>SqlStatisticsPublisherTests\StatisticsPublisherProviderRuntime.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\RelationalStoreTests.cs">
-      <Link>StorageTests\RelationalStoreTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\AdotNetProviderFunctionalityTests.cs">
-      <Link>StorageTests\Relational\AdotNetProviderFunctionalityTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\CommonFixture.cs">
-      <Link>StorageTests\Relational\CommonFixture.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\ConstantHasher.cs">
-      <Link>StorageTests\Relational\ConstantHasher.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\MySqlStorageTests.cs">
-      <Link>StorageTests\Relational\MySqlStorageTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\RelationalStorageTests.cs">
-      <Link>StorageTests\Relational\RelationalStorageTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\SqlServerStorageTests.cs">
-      <Link>StorageTests\Relational\SqlServerStorageTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\StreamingTestRelationalStoragePicker.cs">
-      <Link>StorageTests\Relational\StreamingTestRelationalStoragePicker.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\StorageTests\Relational\TestEnvironmentInvariant.cs">
-      <Link>StorageTests\Relational\TestEnvironmentInvariant.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\LivenessTests.cs">
-      <Link>LivenessTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\MySqlMembershipTableTests.cs">
-      <Link>MySqlMembershipTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\MySqlRemindersTableTests.cs">
-      <Link>MySqlRemindersTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\PostgreSqlMembershipTableTests.cs">
-      <Link>PostgreSqlMembershipTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\PostgreSqlRemindersTableTests.cs">
-      <Link>PostgreSqlRemindersTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\ReminderTests_SqlServer.cs">
-      <Link>ReminderTests_SqlServer.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlServerMembershipTableTests.cs">
-      <Link>SqlServerMembershipTableTests.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\test\TesterSQLUtils\SqlServerRemindersTableTests.cs">
-      <Link>SqlServerRemindersTableTests.cs</Link>
-    </Compile>
+   <ItemGroup>
+    <Compile Include="..\..\..\test\TesterSQLUtils\**\*.cs" Exclude="
+    ..\..\..\test\TesterSQLUtils\obj\**\*.cs;
+    ..\..\..\test\TesterSQLUtils\bin\**\*.cs;
+    ..\..\..\test\TesterSQLUtils\Properties\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\..\src\OrleansSQLUtils\CreateOrleansTables_MySql.sql">
